Create a BaseTest class and reuse it
